operadores aritmeticos - finish

+++ src/main/java/academy/devdojo/maratonajava/introducao/Aula04Operadores.java
@@ -2,9 +2,10 @@ package academy.devdojo.maratonajava.introducao;
 
 public class Aula04Operadores {
     public static void main(String[] args) {
+        // + - / *
         int numero1 = 10;
         int numero2 = 20;
-        int resultado = numero1 / numero2;
+        double resultado = numero1 / (double) numero2;
         System.out.println(resultado);
